// src/types.ts
export interface ApiKeyCreds {
	key: string;
	secret: string;
	passphrase: string;
}
// ...
export enum OrderType {
	GTC = "GTC",
	FOK = "FOK",
	GTD = "GTD",
	FAK = "FAK",
}
// ...
export enum Chain {
	POLYGON = 137,
	AMOY = 80002,
}